rename java module
