+++ Sources/SwiftJavaRuntimeSupport/swift-java.config
@@ -4,6 +4,6 @@
     "org.swift.swiftkit.core.SwiftInstance" : "JavaSwiftInstance",
     "org.swift.swiftkit.core.JNISwiftInstance" : "JavaJNISwiftInstance",
     "org.swift.swiftkit.core.SwiftArena" : "JavaSwiftArena",
-    "org.swift.swiftkit.compose.SwiftObserverCallback" : "JavaSwiftObserverCallback"
+    "org.swift.swiftkit.observable.SwiftObserverCallback" : "JavaSwiftObserverCallback"
   }
 }
