+++ SwiftKitObservable/src/main/kotlin/org/swift/swiftkit/observable/RememberSwiftObservable.kt
@@ -12,7 +12,7 @@
 //
 //===----------------------------------------------------------------------===//
 
-package org.swift.swiftkit.compose
+package org.swift.swiftkit.observable
 
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.DisposableEffect
